cli: migrate to optparse API
This allows us to do less usage message printing ourselves, leaving it
up to the option parsing code instead. Plus we gain the advantage of
type checking the "--profile" option.

Drop the website information from the version number as it doesn't seem
to fit. Finally, start the option help messages with a lower case
letter, which is more in line with common unix commands.