main: Use a gtk.VolumeButton instead of a custom one
Using a gtk.VolumeButton significantly cuts down on the lines of code
needed to support this one button. We can remove callbacks to hide the
scrollwindow, and gain the advantage that it actually hides whenever a
mouse click is made outside the scrollwindow.

Finally, we can simply emit the scroll event from the egg status tray
directly to the button and have it work correctly.